refactor: merge utils into single patterns module
